Reorganized the package structure
