+++ AndroidStudioProject/app/src/main/java/com/fitnesstracker/database/entities/FoodDiaryEntry.java
@@ -1,4 +1,4 @@
-package com.fitnesstracker.database;
+package com.fitnesstracker.database.entities;
 
 import androidx.room.ColumnInfo;
 import androidx.room.Entity;
