+++ AndroidStudioProject/app/src/main/res/layout/fragment_add_exercise.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.AddExerciseFragment">
+    tools:context=".ui.fragments.AddExerciseFragment">
 
     <TextView
         android:layout_width="match_parent"
