+++ AndroidStudioProject/app/src/main/res/layout/fragment_add_food.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.AddFoodFragment">
+    tools:context=".ui.fragments.AddFoodFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
