+++ AndroidStudioProject/app/src/main/res/layout/fragment_diary.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.DiaryFragment">
+    tools:context=".ui.fragments.DiaryFragment">
 
     <com.google.android.material.floatingactionbutton.FloatingActionButton
         android:id="@+id/add_diary_entry_fab"
