+++ AndroidStudioProject/app/src/main/res/layout/fragment_exercise.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.ExerciseFragment">
+    tools:context=".ui.fragments.ExerciseFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
