+++ AndroidStudioProject/app/src/main/res/layout/fragment_food.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.FoodFragment">
+    tools:context=".ui.fragments.FoodFragment">
 
     <com.google.android.material.floatingactionbutton.FloatingActionButton
         android:id="@+id/add_food_fab"
