+++ AndroidStudioProject/app/src/main/res/navigation/nav_fitnesstracker.xml
@@ -7,17 +7,17 @@
 
     <fragment
         android:id="@+id/diaryFragment"
-        android:name="com.fitnesstracker.ui.DiaryFragment"
+        android:name="com.fitnesstracker.ui.fragments.DiaryFragment"
         android:label="fragment_diary"
         tools:layout="@layout/fragment_diary" />
     <fragment
         android:id="@+id/exerciseFragment"
-        android:name="com.fitnesstracker.ui.ExerciseFragment"
+        android:name="com.fitnesstracker.ui.fragments.ExerciseFragment"
         android:label="fragment_exercise"
         tools:layout="@layout/fragment_exercise" />
     <fragment
         android:id="@+id/foodFragment"
-        android:name="com.fitnesstracker.ui.FoodFragment"
+        android:name="com.fitnesstracker.ui.fragments.FoodFragment"
         android:label="fragment_food"
         tools:layout="@layout/fragment_food">
         <action
@@ -26,12 +26,12 @@
     </fragment>
     <fragment
         android:id="@+id/addFoodFragment"
-        android:name="com.fitnesstracker.ui.AddFoodFragment"
+        android:name="com.fitnesstracker.ui.fragments.AddFoodFragment"
         android:label="fragment_add_food"
         tools:layout="@layout/fragment_add_food" />
     <fragment
         android:id="@+id/homeFragment"
-        android:name="com.fitnesstracker.ui.HomeFragment"
+        android:name="com.fitnesstracker.ui.fragments.HomeFragment"
         android:label="fragment_home"
         tools:layout="@layout/fragment_home" />
 </navigation>
